correcion ejercicio  ∾ prgtest resultat 31_37: FALLA

+++ .introprg/tmp/test/NotaMitja.java
@@ -3,14 +3,14 @@
 public class NotaMitja {
     public static void main(String[]args){
      // creem la variable on guardarem el resultat de sumar
-        double suma = 0;   // inicialment no hem sumat res i per tant és 0
-        double promedio =  0.0  ;
-        double valor = 0;
-        double contador= 1;
+        int suma = 0;   // inicialment no hem sumat res i per tant és 0
+        int promedio = 0;
+        int valor = 0;
+        int contador= 1;
         
         // demana el primer valor
         System.out.println("Introdueix un valor");       
-        valor = Double.parseDouble(Entrada.readLine());   
+        valor = Integer.parseInt(Entrada.readLine());   
         
        if(valor < 0 || valor > 100){
                 System.out.println("Cap nota vàlida introduïda");  
@@ -19,7 +19,7 @@ public class NotaMitja {
         // ves sumant mentre el valor sigui positiu
         while (valor >= 0 && valor <= 100) {
             System.out.println("Introdueix un valor");
-                valor = Double.parseDouble(Entrada.readLine());
+                valor = Integer.parseInt(Entrada.readLine());
             if (valor >= 0 && valor <= 100){
                 suma = suma + valor;
                 contador= contador + 1;
